LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/background-size-expected.txt


PASS Can set 'background-size' to CSS-wide keywords: initial
PASS Can set 'background-size' to CSS-wide keywords: inherit
PASS Can set 'background-size' to CSS-wide keywords: unset
PASS Can set 'background-size' to CSS-wide keywords: revert
FAIL Can set 'background-size' to var() references:  var(--A) assert_equals: expected 2 but got 1
FAIL Can set 'background-size' to a length: 0px assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'background-size' to a length: -3.14em assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'background-size' to a length: 3.14cm assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'background-size' to a length: calc(0px + 0em) assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSStyleValue]"
FAIL Can set 'background-size' to a percent: 0% assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'background-size' to a percent: -3.14% assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'background-size' to a percent: 3.14% assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'background-size' to a percent: calc(0% + 0%) assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSStyleValue]"
FAIL Can set 'background-size' to the 'auto' keyword: auto assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
PASS Can set 'background-size' to the 'cover' keyword: cover
PASS Can set 'background-size' to the 'contain' keyword: contain
PASS Setting 'background-size' to a time throws TypeError
PASS Setting 'background-size' to an angle throws TypeError
PASS Setting 'background-size' to a flexible length throws TypeError
FAIL Setting 'background-size' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'background-size' to a URL throws TypeError
PASS Setting 'background-size' to a transform throws TypeError
PASS 'background-size' does not supported '200px 100px'
